Refactor hospital list and table
